test(dashboard): fold EmptyState DOM tests into feedback-state.test.ts, drop stale duplicate
empty-state.test.ts was a drift-prone duplicate: it tested EmptyState (which
lives in feedback-state.ts) under a misleading filename and a blanket
@ts-nocheck pragma that hid a real type error (h(EmptyState) needs the props
arg). Its 7 DOM-shape assertions (role=status, message, icon, children>message,
compact padding, custom class, action slot) are genuine coverage not present in
the canonical feedback-state.test.ts (which covers EmptyState metadata only).

Fold those 7 tests into feedback-state.test.ts as an 'EmptyState DOM' block
(fixing h(EmptyState) -> h(EmptyState, {})), and delete empty-state.test.ts so
one component is covered by one test file. No @ts-nocheck; tsc clean.
